# array_get_element function arguments from float to int cause it's not a position on screen

+++ headers/UI/array.h
@@ -16,7 +16,7 @@ typedef struct Array
 int array_set_fields(Array *array, UI_Element **elements, int element_count, int max_per_line, int padding, int elements_anchor);
 Array *array_init(Array *array);
 Array *array_set_elements_position(Array *array);
-UI_Element *array_get_element(Array *array, float x, float y);
+UI_Element *array_get_element(Array *array, int x, int y);
 int array_get_width(Array *array);
 int array_get_height(Array *array);
